extract module evaluators (#16)
* rename module `evaluator` to `evaluators`

* Update README.md

* move EvalutorBase class into Evaluators

* Update llm_evaluator.py

* rename corresponding test module

* rename test modules

* update doc

* Update my_first_agent.ipynb
