+++ council/evaluators/__init__.py
@@ -1,4 +1,5 @@
 """This package provides ready to use evaluators"""
 
+from .evaluator_base import EvaluatorBase
 from .basic_evaluator import BasicEvaluator
 from .llm_evaluator import LLMEvaluator
