+++ council/evaluators/basic_evaluator.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from council.core import AgentContext, Budget
-from council.core.evaluator_base import EvaluatorBase
+from council.evaluators.evaluator_base import EvaluatorBase
 from council.core.execution_context import (
     ScoredAgentMessage,
     SkillSuccessMessage,
