--- docs/source/reference/core/evaluator_base.rst
@@ -1,4 +1,0 @@
-EvaluatorBase
--------------
-
-.. autoclass:: council.core.EvaluatorBase
