--- docs/source/reference/evaluator.rst
@@ -1,14 +1,0 @@
-Evaluator
-=========
-
-.. automodule:: council.evaluator
-
-.. autoclasstree:: council.evaluator
-    :full:
-    :namespace: council
-
-.. toctree::
-    :maxdepth: 1
-    :glob:
-
-    evaluator/*
