--- docs/source/reference/evaluator/basic_evaluator.rst
@@ -1,6 +1,0 @@
-BasicEvaluator
---------------
-
-.. autoclasstree:: council.evaluator.BasicEvaluator
-
-.. autoclass:: council.evaluator.BasicEvaluator
