--- docs/source/reference/evaluator/llm_evaluator.rst
@@ -1,6 +1,0 @@
-LLMEvaluator
-------------
-
-.. autoclasstree:: council.evaluator.LLMEvaluator
-
-.. autoclass:: council.evaluator.LLMEvaluator
