+++ docs/source/reference/evaluators.rst
@@ -1,0 +1,14 @@
+Evaluators
+----------
+
+.. automodule:: council.evaluators
+
+.. autoclasstree:: council.evaluators
+    :full:
+    :namespace: council
+
+.. toctree::
+    :maxdepth: 1
+    :glob:
+
+    evaluators/*
