+++ docs/source/reference/evaluators/basic_evaluator.rst
@@ -1,0 +1,6 @@
+BasicEvaluator
+--------------
+
+.. autoclasstree:: council.evaluators.BasicEvaluator
+
+.. autoclass:: council.evaluators.BasicEvaluator
