+++ docs/source/reference/evaluators/evaluator_base.rst
@@ -1,0 +1,4 @@
+EvaluatorBase
+-------------
+
+.. autoclass:: council.evaluators.EvaluatorBase
