+++ docs/source/reference/evaluators/llm_evaluator.rst
@@ -1,0 +1,6 @@
+LLMEvaluator
+------------
+
+.. autoclasstree:: council.evaluators.LLMEvaluator
+
+.. autoclass:: council.evaluators.LLMEvaluator
