+++ tests/unit/agents/test_agent.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 from council.agents import Agent
 from council.controllers import BasicController
-from council.evaluator import BasicEvaluator
+from council.evaluators import BasicEvaluator
 from council.mocks import MockSkill
